changes to blpAuthenticate
