+++ R/RcppExports.R
@@ -1,6 +1,10 @@
 # This file was generated by Rcpp::compileAttributes
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+authenticate_Impl <- function(conn_, uuid_, ip_address_) {
+    .Call('Rblpapi_authenticate_Impl', PACKAGE = 'Rblpapi', conn_, uuid_, ip_address_)
+}
+
 bdh_Impl <- function(conn_, securities, fields, start_date_, end_date_, options_, identity_) {
     .Call('Rblpapi_bdh_Impl', PACKAGE = 'Rblpapi', conn_, securities, fields, start_date_, end_date_, options_, identity_)
 }
